requirements.txt
azure-cognitiveservices-speech==1.30.0
python-dotenv
gradio